+++ paymentsheet-example/src/main/java/com/stripe/android/paymentsheet/example/playground/activity/PaymentSheetPlaygroundActivity.kt
@@ -882,7 +882,7 @@ class PaymentSheetPlaygroundActivity : AppCompatActivity() {
                 /**
                  * Modify this list if you want to change the countries displayed in the playground.
                  */
-                country.code.value in setOf("US", "GB", "AU", "FR", "IN", "SG", "MY")
+                country.code.value in setOf("US", "GB", "AU", "FR", "IN", "SG", "MY", "MX")
             }.map { country ->
                 /**
                  * Modify this statement to change the default currency associated with each
@@ -910,6 +910,9 @@ class PaymentSheetPlaygroundActivity : AppCompatActivity() {
                     "MY" -> {
                         country to "MYR"
                     }
+                    "MX" -> {
+                        country to "MXN"
+                    }
                     else -> {
                         country to "USD"
                     }
@@ -918,6 +921,6 @@ class PaymentSheetPlaygroundActivity : AppCompatActivity() {
 
         // List was created from: https://stripe.com/docs/currencies
         /** Modify this list if you want to change the currencies displayed in the playground **/
-        private val stripeSupportedCurrencies = listOf("AUD", "EUR", "GBP", "USD", "INR", "PLN", "SGD", "MYR")
+        private val stripeSupportedCurrencies = listOf("AUD", "EUR", "GBP", "USD", "INR", "PLN", "SGD", "MYR", "MXN")
     }
 }