+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## XX.XX.XX - 2023-XX-XX
 
+### PaymentSheet
+
+* [ADDED] PaymentSheet now supports the following payment methods for PaymentIntents:
+  * [7281](https://github.com/stripe/stripe-android/pull/7281) OXXO
+
 ## 20.30.1 - 2023-09-11
 
 ### Financial Connections